Refactor homepage too
