+++ src/app/app.module.ts
@@ -4,9 +4,9 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppComponent } from './app.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { FormsModule } from '@angular/forms';
-import { TopNavbarComponent } from './top-navbar/top-navbar.component';
+import { TopNavbarComponent } from './home/top-navbar/top-navbar.component';
 import { AppRoutingModule } from './app-routing.module';
-import { HomepageComponent } from './homepage/homepage.component';
+import { HomepageComponent } from './home/homepage/homepage.component';
 import { ProjectsComponent } from './base/projects/projects.component';
 import { BlogComponent } from './base/blog/blog.component';
 import { AboutComponent } from './base/about/about.component';
