+++ src/app/markdown/article-md-display/article-md-display.component.html
@@ -1,5 +1,5 @@
 <div *ngIf="article">
     <div class="container">
-        <ngx-math-md katex [src]="article.path"></ngx-math-md >
+        <ngx-math-md katex [src]="article.path"></ngx-math-md>
     </div>
 </div>
