+++ src/app/markdown/article-md-display/article-md-display.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { MdService } from '../../md.service';
+import { MdService } from '../../service/md.service';
 import { Article } from '../../model/Article';
 
 @Component({
